// src/app/interfaces/simRegisteration.model.ts
export interface SimRegistrationData {
    firstName: string;
    lastName: string;
    birthDate: string;
    state: string;
    province: string;
    city: string;
    gender: string;
    phoneNumber: string;
    email: string;
    idNumber: string;
    idType: string;
    address: string;
    postalCode: string;
    emergencyContact: string;
    emergencyPhone: string;
}